requirements.txt
fastapi
uvicorn
strawberry-graphql
asyncpg
psycopg2-binary
python-dotenv
